fix: restructuring code
