+++ app/(web)/user/[uuid]/page.js
@@ -1,42 +1,31 @@
-import { getUserInfoById } from '../edit/actions/userService'
+import { getDataFromTable } from '../upload/actions/getSoftwares'
+import { getImageInfoById } from './actions/getImage'
+import EditImage from './components/EditImage'
 
 export default async function Page({ params }) {
   const { uuid } = params
+  console.log(uuid, 7)
+  if (!uuid) {
+    return <div>Invalid ID</div>
+  }
+  const { error, data: imageInfo } = await getImageInfoById(uuid)
 
-  let userInfo = null
-  let error = null
-
-  try {
-    userInfo = await getUserInfoById(uuid)
-  } catch (e) {
-    error = 'Error fetching user details: ' + e.message
+  if (!imageInfo) {
+    return notFound()
   }
 
   if (error) {
     return <div className="text-red-500">{error}</div>
   }
 
-  if (!userInfo) {
-    return <div>Loading...</div>
-  }
-
-  const { username, avatar_file_exists, cover_file_exists, email } = userInfo
+  const softwareOptions = (await getDataFromTable('softwares')).data
+  const tagsOptions = (await getDataFromTable('tags')).data
 
   return (
-    <div>
-      <h1>User Details</h1>
-      <p>
-        <strong>Username:</strong> {username}
-      </p>
-      <p>
-        <strong>Email:</strong> {email}
-      </p>
-      <p>
-        <strong>Avatar Exists:</strong> {avatar_file_exists ? 'Yes' : 'No'}
-      </p>
-      <p>
-        <strong>Cover Exists:</strong> {cover_file_exists ? 'Yes' : 'No'}
-      </p>
-    </div>
+    <EditImage
+      imageInfo={imageInfo}
+      softwareOptions={softwareOptions}
+      tagsOptions={tagsOptions}
+    />
   )
 }
